refactor: 장바구니 페이지 할인율 계산, 재고 계산 로직을 util로 분리
